Main board schematic.
This project will ultimately be made from two separate PCBs. The main
board contains the microcontroller and RTC, while the display board
(yet to be designed) will contain the display LEDs and their associated
driver chips. The communication between the two boards is via an SPI bus,
augmented with a ~RESET line so that the display drivers can reset along
with everything else.