Correctly configure the UART.
Some mis-reading of the datasheet led to a complete misunderstanding of
the structure of the USART0_CFG register, causing the UART to be
configured for 2 stop bits, 7-bit data length and even parity. Despite
how wrong this was, it ended up generating a bunch of bytes correctly,
making it hard at first to recognize the problem.

The new setting is, as the comment originally suggested, 1 stop bit,
8-bit data length and no parity bit.