Fleshed out scheduler.
It now basically works and supports timer events. It's using the
SysTick timer, so we run through the event loop at least once every
10ms.

Ideally we would set the timer to occur at the time of the next timer
in the list, but so far I couldn't find a way to measure the time
elapsed while the processor is asleep and so waking up every 10ms is
necessary to keep track of the (approximate) passage of time. Hopefully
a better solution will present itself later, but this will work for
now.