Button input interface.
Interacts with the scheduler to allow tasks to do blocking reads of the
direction buttons, with them becoming runnable again only after we get
a debounced press down.

This initial implementation has a bug where pressing a second button when
one is already pressed will cause a spurious press event when one of the
buttons is released. This can be refined later by keeping track of the
previous state so that only true differences are reported.