Prototype display driver.
The display is built from "display elements", which are defined by a
row generator function and an integer argument to provide to it. Each
element can generate zero or more display rows (which will actually be
columns on the final display).

Eventually each UI state will have its own set of display elements that
together define what is displayed on-screen in that state. The display
manager will also eventually be able to show partial renderings of two
different states on-screen in order to implement the inter-state
transitions.

For now, since the "real" display board isn't built yet, we produce a
simulated version of the display on the UART.