Flip the animation directions.
Previously the whole-screen transitions went in opposition to the
direction arrows, as if moving the "camera" over a set of items, while
the hour-to-second transition was the opposite.

For consistency, now the animation moves in the direction of the arrow
pressed, as if the user is moving a "canvas" that all of the UI elements
are on.